beans have less logic inside. Move resource name validator to common package.
