+++ galasa-parent/dev.galasa.framework.api.common/src/main/java/dev/galasa/framework/api/common/resources/beans/GalasaPropertyData.java
@@ -1,0 +1,13 @@
+package dev.galasa.framework.api.common.resources.beans;
+
+public class GalasaPropertyData {
+    private String value;
+
+    public GalasaPropertyData (String propertyValue){
+        this.value = propertyValue;
+    }
+
+    public String getValue() {
+        return this.value;
+    }
+}
