+++ galasa-parent/dev.galasa.framework.api.common/src/test/java/dev/galasa/framework/api/common/resources/TestGalasaProperty.java
@@ -8,6 +8,7 @@ package dev.galasa.framework.api.common.resources;
 import org.junit.Test;
 
 import dev.galasa.framework.api.common.InternalServletException;
+import dev.galasa.framework.api.common.resources.beans.GalasaProperty;
 
 import static org.assertj.core.api.Assertions.*;
