+++ galasa-parent/dev.galasa.framework.api.cps/bnd.bnd
@@ -1,7 +1,7 @@
 -snapshot ${tstamp}
 Bundle-Name: Galasa API Configuration Property Store microservices
 Export-Package: dev.galasa.framework.api.cps*;
-Import-Package: dev.galasa.framework.api.common, \
+Import-Package: dev.galasa.framework.api.common*, \
                 dev.galasa.framework.api.authentication, \
                 *
