+++ galasa-parent/dev.galasa.framework.openapi.generator/src/main/java/BeanGenerator.java
@@ -1,0 +1,5 @@
+package dev.galasa.framework.openapi.generator;
+
+public class BeanGenerator {
+
+}
